Move entry-reader pckage
