+++ internal/scheduler/testdata/stop.yaml
